Add new Node System under SytemEx.Model - future replace old Node under SystemEx.Collections.Generic.
Remove bug in Array

+++ Base/NumberRangeIterrator.cs
@@ -67,7 +67,7 @@ namespace SystemEx {
         /// <summary>
         /// Moves the iterator N step forward
         /// </summary>
-        public void Forward (int i) {
+        public void Forward ( long i ) {
             var n = i;
             while ( n > 0 ) {
                 --n;
