+++ Collections/Generic/PriorityQueue.cs
@@ -83,7 +83,7 @@ namespace SystemEx.Collections.Generic {
         /// <summary>
         /// Moves the iterator N step forward
         /// </summary>
-        public void Forward ( int i ) {
+        public void Forward ( long i ) {
             var n = i;
             while ( n > 0 ) {
                 --n;
